Add new Japan endpoint for GPT4 and refactoring
