+++ utils/app/const.ts
@@ -43,6 +43,7 @@ export const AZURE_OPENAI_KEYS = [
   process.env.AZURE_OPENAI_KEY_3,
 ];
 
+// To be enabled, currently we goes with this order of region regardless of request origin
 export const ENDPOINT_TRAFFIC_DISTRIBUTION = [
   0.75, // Japan East
   0.10, // Central US
@@ -51,11 +52,13 @@ export const ENDPOINT_TRAFFIC_DISTRIBUTION = [
 ];
 
 export const AZURE_OPENAI_GPT_4_ENDPOINTS = [
+  process.env.AZURE_OPENAI_GPT_4_ENDPOINT_0,
   process.env.AZURE_OPENAI_GPT_4_ENDPOINT_1,
   process.env.AZURE_OPENAI_GPT_4_ENDPOINT_2,
 ];
 
 export const AZURE_OPENAI_GPT_4_KEYS = [
+  process.env.AZURE_OPENAI_GPT_4_KEY_0,
   process.env.AZURE_OPENAI_GPT_4_KEY_1,
   process.env.AZURE_OPENAI_GPT_4_KEY_2,
 ];
